+++ CHANGELOG.md
@@ -13,6 +13,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Sync: `select_all` option for custom format groups to include all CFs regardless of default
   status (#722)
+- Sync: Warning when `custom_format_groups.skip` contains trash_ids that don't match any known CF
+  group
 
 ### Fixed